requirements.txt
dashscope==1.19.2
fastapi==0.111.0
replicate==0.26.0
Requests==2.32.3
uvicorn==0.30.1
alibabacloud-credentials==0.3.3
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.2
alibabacloud-sts20150401==1.1.4
alibabacloud-tea==0.3.6
alibabacloud-tea-openapi==0.3.9
alibabacloud-tea-util==0.3.12
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
oss2~=2.18.5
PyJWT==2.8.0
SQLAlchemy==2.0.31
pymysql==1.1.1
pillow~=10.3.0